+++ client/src/components/ErrorMessage.tsx
@@ -5,6 +5,8 @@ import { CloseS2 } from 'plaid-threads/Icons/CloseS2';
 
 import useErrors from '../services/errors';
 
+// link errors that require a prompt for the enduser
+
 export default function ErrorMessage() {
   const [show, setShow] = useState(false);
   const [message, setMessage] = useState('');
