+++ client/src/components/index.ts
@@ -1,7 +1,6 @@
-export { default as AccountBalanceCheck } from './AccountBalanceCheck';
+export { default as Transfers } from './Transfers';
 export { default as AddUserForm } from './AddUserForm';
 export { default as Banner } from './Banner';
-export { default as DuplicateItemToastMessage } from './DuplicateItemToast';
 export { default as ItemInfo } from './ItemInfo';
 export { default as UserPage } from './UserPage';
 export { default as Landing } from './Landing';
@@ -13,6 +12,6 @@ export { default as UserList } from './UserList';
 export { default as OAuthLink } from './OAuthLink';
 export { default as Login } from './Login';
 export { default as ErrorMessage } from './ErrorMessage';
-export { default as TransferFunds } from './TransferFunds';
+export { default as TransferForm } from './TransferForm';
 export { default as ConfirmIdentity } from './ConfirmIdentity';
-export { default as MainAccount } from './MainAccount';
+export { default as PatternAccount } from './PatternAccount';
